Increase projectile speed slightly across the board.
This bumps arrows from 75 to 100, slingers to 90 and javelins from 62.5
to 70. It's both slightly more realistic and helps with dancing, since
the problem is the ratio of unit speed vs projectile speed.

Further increases would reduce dancing, but also make it harder to see
projectiles which isn't desirable.

Refs #5106

Differential Revision: https://code.wildfiregames.com/D3179
This was SVN commit r24458.