[gameplay-a24] Lower cost of briton special tech for grain gather rate
The technology introduced in 586b045d09 for gauls is a bit overprised.
This is lowering the price to the original cost of rotarymill and
lowering gain a bit more because the tech is not just gobal but
available in phase earlier. Also it is "almost" half on cost and effect
of the current phase II tech which is available for everyone.

Differential Revision: D3253
Reviewed by: borg-
This was SVN commit r24460.