Regroup special technologies.
Some techs are shared between multiple civs and having them in a folder
that is named differently (successors/hellenes) can be confusing.
This patch deprecates the subfolders and moves all techs in the main
folder.

(Also did svn propset on *.json to not have to do it for every changed
template.)

Patch by: @Nescio
Differential revision: D888
Comment by: @fatherbushido
This was SVN commit r24558.