Fix PetraAI not building many houses.
Previous to 7f77cf2f3e the AI got its population bonus per entity from
the Cost component, which every entity happened to have.
Thereafter, the population bonus was handled by a separate component,
which was not always present leading to `(void 0)` additions to the pop
cap and PetraAI not constructing houses.

Differential revision: D3360
Reviewed by: @Angen, @wraitii
This was SVN commit r24613.