[Gameplay A24] - Reduce ranged soldier resistance.
A frequent complaint is that ranged troops are too effective in 0 A.D.
They're supposed to die quickly, however, they get the same resistance
increases as melee troops. This limits the advanced and elite resistance
bonus to melee only (thus also excluding healers) and reduces the
resistance of ranged champions and heroes by two levels, to stay in pace
with their elite counterparts.
As a consequence, nothing changes for basic ranged troops, but advanced
are about 11% easier to kill, and elite, champion, and heroes about 23%.

Patch by: @Nescio
Differential revision: D3369
Reviewed by: @wraitii
Comment by: @borg-
This was SVN commit r24636.