Improve detection of empty notification messages
In the case that a player started or aborted researching a phase
technology, a
blank line was being added to the chat history of the player in question
and
any mutual ally or observer, where the player, ally, or observer had
their
`gui.session.notifications.phase` setting at the default value of
"completed".

This was happening because in the above case the returned value of
`ChatMessageFormatSimulation.phase` class's `.parse()` method is `{text:
""}`.


Related a3eccc043d
Fixes #5921

This was SVN commit r24671.