[gameplay-a24] Tweak elephant stable building time
As pointed out by Edwarf in the forum post, the elephant stable is
rather expensive compared to the arsenal (both are city-phase
structures).
This patch keeps the elephant stable's resource costs unchanged (it is
visually larger than the other three military structures), but lowers
the building time from 240 s to 180 s.

Differential revision: D3515
Reviewed by: @borg-
Patch by: @Nescio
This was SVN commit r24850.