Allow capture when falling back to "WalkAndFight" in COMBAT.APPROACHING
COMBAT.APPROACHING will fallback to WalkAndFight if the order is forced
and the movement fails (happens e.g. on target death).
That behaviour was added in 5568bd4c16 (and tweaked in d9c6879450), to
make the unit move to the general position asked by the player as a
fallback.
However, it did not allow capture, which meant units would occasionally
start attacking structures, instead of capturing (the default).
This makes that fallback allow capture in all cases, which is generally
more expected than the reverse.

Fixes #6003

Differential Revision: https://code.wildfiregames.com/D3548
This was SVN commit r24866.