Fix no-formation orders after several formation orders.
When giving several queued formation-walk orders, then a queued
non-formation order (such as gathering), the gather order would be
silently dropped. This did not happen if only one formation-walk order
was given, which is why this issue was not noticed in 59d0885d68

To fix it, explicitly only drop pre-"leave formation" orders in the
queue for formation members.

Reported By: faction02
Differential Revision: https://code.wildfiregames.com/D3550
This was SVN commit r24869.