Remove territory root from fortresses.
Reverts D1762 / 014c2922f2.
Because any territory connected to a root does not decay, this made it
possible to anchor CC territories by adding a fortress somewhere, thus
making it much, much harder to push against an enemy territory.
This led to a much more turtley gameplay and overall made the game
worse.

For now, I do not reintroduce the larger territory range (which was
reduced in 014c2922f2).

Differential Revision: https://code.wildfiregames.com/D3989
This was SVN commit r25697.