RM generation: improve metal mines generation
Introduces a new helper to place mines. use it on 'Best for MP' maps for
stone and metal.

The amount of metal will scale with the # of players, and will place
slightly more mines overall for fairer distribution.

This should make 1v1 and 4v4 more playable on these maps.

Differential Revision: https://code.wildfiregames.com/D4148
This was SVN commit r25772.