Fix lobby dgame details panel with incompatible mods.
There was an issue in 07e44a75a1 that made the game details mostly red
with incompatible mods.
This fixes that: now only the 'Mods' title is red, and I added specific
text to highlight why it's colored that way.

Based on a patch by: Langbart
Differential Revision: https://code.wildfiregames.com/D4153
This was SVN commit r25778.